requirements.txt
aiogram==2.21
loguru==0.6.0
tortoise-orm==0.19.1  # Database